geseca: switch to float

+++ src/ScoreCalculation.cpp
@@ -1,6 +1,6 @@
 #include "ScoreCalculation.h"
 
-double getVarianceEstimator(const std::vector<double> & elems){
+double getVarianceEstimator(const std::vector<float> & elems){
     double varHat = 0.0;
     for (double elem : elems){
         varHat += elem * elem;
@@ -10,11 +10,11 @@ double getVarianceEstimator(const std::vector<double> & elems){
 }
 
 
-std::vector<double> getProfile(const std::vector<double> & E,
+std::vector<float> getProfile(const std::vector<float> & E,
                                const std::vector<unsigned> & indexes,
                                unsigned m){
     // unsigned m = E[0].size(); // number or rows must be nonzero
-    std::vector<double> profile(m);
+    std::vector<float> profile(m);
 
     for (unsigned j = 0; j < m; j++){
         double value = 0.0;
@@ -27,15 +27,15 @@ std::vector<double> getProfile(const std::vector<double> & E,
 }
 
 
-double getScore(const std::vector<double> & profile){
+double getScore(const std::vector<float> & profile){
     return(getVarianceEstimator(profile));
 }
 
 
 
-void adjustProfile(const std::vector<double> & E,
-                                  const std::vector<double> & profile,
-                                  std::vector<double> & newProfile,
+void adjustProfile(const std::vector<float> & E,
+                                  const std::vector<float> & profile,
+                                  std::vector<float> & newProfile,
                                   unsigned idNew, unsigned idOld,
                                   unsigned m){
     for (unsigned i = 0; i < newProfile.size(); i++){
